requirements.txt
et_xmlfile==2.0.0
openpyxl==3.1.5
psycopg2==2.9.10
python-dotenv==1.1.0
